requirements.txt
python-dotenv==0.17.1
requests==2.32.5
selenium==4.40.0
webdriver-manager==4.0.2
